fix: Remove first and last name from user, move requests to different file
